Revert "Refactor importers to separate vocabulary importers from base"
This reverts commit 668535c7071e0a163797903eea37b48c2805fb3c.
