+++ ruby/lib/goi/importer.rb
@@ -1,3 +1,4 @@
-require_relative 'importer/core'
-require_relative 'importer/vocabulary'
-require_relative 'importer/grammar'
+require_relative 'importer/config'
+require_relative 'importer/base_importer'
+require_relative 'importer/google_sheet_importer'
+require_relative 'importer/sequel_importer'
