+++ ruby/lib/goi/importer/base_importer.rb
@@ -1,0 +1,17 @@
+module Goi
+  module Importer
+    class BaseImporter
+
+      def initialize(config:)
+        @config = config
+      end
+
+      attr_reader :config
+
+      def inport
+        # Do Nothing
+      end
+
+    end
+  end
+end
