--- ruby/lib/goi/importer/core.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'core/config'
-require_relative 'core/base_importer'
