--- ruby/lib/goi/importer/grammar.rb
@@ -1,2 +1,0 @@
-# frozen_string_literal: true
-
