--- ruby/lib/goi/importer/vocabulary.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'vocabulary/google_sheet_importer'
-require_relative 'vocabulary/sequel_importer'
