+++ ruby/sig/lib/goi/importer/base_importer.rbs
@@ -1,0 +1,11 @@
+module Goi
+  module Importer
+    class BaseImporter
+      def initialize: (config: Config) -> void
+
+      attr_reader config: Config
+
+      def parse: () -> [Model::Vocabulary::Linkages]
+    end
+  end
+end
