+++ ruby/sig/lib/goi/importer/config.rbs
@@ -1,0 +1,15 @@
+module Goi
+  module Importer
+    class Config
+
+      type db_config = {database: String, user: String, password: String, host: String}
+
+      def initialize: (file_pathname: Pathname?, db_config: db_config?) -> void
+
+      attr_reader file_pathname: Pathname?
+
+      attr_reader db_config: db_config?
+
+    end
+  end
+end
