--- ruby/sig/lib/goi/importer/core/base_importer.rbs
@@ -1,13 +1,0 @@
-module Goi
-  module Importer
-    module Core
-      class BaseImporter
-        def initialize: (config: Config) -> void
-
-        attr_reader config: Config
-
-        def parse: () -> [Model::Vocabulary::Linkages]
-      end
-    end
-  end
-end
